feat: #247 remove web wetting from security
