+++ src/main/java/com/umc/linkyou/config/security/RedisAuthorizationRequestRepository.java
@@ -1,4 +1,4 @@
-package com.umc.linkyou.oauth.web;
+package com.umc.linkyou.config.security;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.servlet.http.HttpServletRequest;
