--- src/main/java/com/umc/linkyou/oauth/mobile/client/NaverTokenClient.java
@@ -1,4 +1,0 @@
-package com.umc.linkyou.oauth.mobile.client;
-
-public class NaverTokenClient {
-}
