--- src/main/java/com/umc/linkyou/oauth/mobile/controller/MobileAuthController.java
@@ -1,4 +1,0 @@
-package com.umc.linkyou.oauth.mobile.controller;
-
-public class MobileAuthController {
-}
