--- src/main/java/com/umc/linkyou/oauth/mobile/dto/MobileLoginResponse.java
@@ -1,4 +1,0 @@
-package com.umc.linkyou.oauth.mobile.dto;
-
-public class MobileLoginResponse {
-}
