--- src/main/java/com/umc/linkyou/oauth/mobile/service/KakaoMobileAuthService.java
@@ -1,4 +1,0 @@
-package com.umc.linkyou.oauth.mobile.service;
-
-public class KakaoMobileAuthService {
-}
