+++ src/main/java/com/umc/linkyou/oauth2/mobile/client/KakaoTokenClient.java
@@ -1,0 +1,4 @@
+package com.umc.linkyou.oauth2.mobile.client;
+
+public class KakaoTokenClient {
+}
