+++ src/main/java/com/umc/linkyou/oauth2/mobile/controller/MobileAuthController.java
@@ -1,0 +1,4 @@
+package com.umc.linkyou.oauth2.mobile.controller;
+
+public class MobileAuthController {
+}
