+++ src/main/java/com/umc/linkyou/oauth2/mobile/dto/MobileLoginRequest.java
@@ -1,0 +1,4 @@
+package com.umc.linkyou.oauth2.mobile.dto;
+
+public class MobileLoginRequest {
+}
