+++ src/main/java/com/umc/linkyou/oauth2/mobile/service/GoogleMobileAuthService.java
@@ -1,0 +1,4 @@
+package com.umc.linkyou.oauth2.mobile.service;
+
+public class GoogleMobileAuthService {
+}
