+++ src/main/java/com/umc/linkyou/oauth2/utils/KakaoUserInfoExtractor.java
@@ -1,4 +1,4 @@
-package com.umc.linkyou.oauth.utils;
+package com.umc.linkyou.oauth2.utils;
 
 import org.springframework.security.oauth2.core.user.OAuth2User;
 import org.springframework.stereotype.Component;
